+++ src/Workspaces/Core/Portable/ExternalAccess/UnitTesting/Api/IUnitTestingIncrementalAnalyzerImplementation.cs
@@ -11,9 +11,9 @@ namespace Microsoft.CodeAnalysis.ExternalAccess.UnitTesting.Api
         Task DocumentOpenAsync(Document document, CancellationToken cancellationToken);
         Task DocumentCloseAsync(Document document, CancellationToken cancellationToken);
         Task DocumentResetAsync(Document document, CancellationToken cancellationToken);
-        Task AnalyzeSyntaxAsync(Document document, UnitTestingInvovationReasonsWrapper reasons, CancellationToken cancellationToken);
-        Task AnalyzeDocumentAsync(Document document, SyntaxNode bodyOpt, UnitTestingInvovationReasonsWrapper reasons, CancellationToken cancellationToken);
-        Task AnalyzeProjectAsync(Project project, bool semanticsChanged, UnitTestingInvovationReasonsWrapper reasons, CancellationToken cancellationToken);
+        Task AnalyzeSyntaxAsync(Document document, UnitTestingInvocationReasonsWrapper reasons, CancellationToken cancellationToken);
+        Task AnalyzeDocumentAsync(Document document, SyntaxNode bodyOpt, UnitTestingInvocationReasonsWrapper reasons, CancellationToken cancellationToken);
+        Task AnalyzeProjectAsync(Project project, bool semanticsChanged, UnitTestingInvocationReasonsWrapper reasons, CancellationToken cancellationToken);
         void RemoveDocument(DocumentId documentId);
         void RemoveProject(ProjectId projectId);
         bool NeedsReanalysisOnOptionChanged(object sender, UnitTestingOptionChangedEventArgsWrapper e);
